# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Threat Infrastructure Mapper (TIM) API",
    "description": "Local-first threat intelligence and infrastructure mapping. Submit one IOC and TIM collects, fingerprints, enriches, pivots and correlates the infrastructure that belongs to the same operation.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/graph/{graph_id}": {
      "get": {
        "tags": [
          "Graph & Clusters"
        ],
        "summary": "Investigation / cluster / asset relationship graph (NetworkX, React Flow format)",
        "operationId": "get_graph_graph__graph_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "graph_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Graph Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "auto",
                "investigation",
                "cluster",
                "asset"
              ],
              "type": "string",
              "default": "auto",
              "title": "Kind"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3,
              "minimum": 1,
              "description": "Neighbourhood depth when kind=asset",
              "default": 1,
              "title": "Depth"
            },
            "description": "Neighbourhood depth when kind=asset"
          },
          {
            "name": "max_nodes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 10,
              "default": 1500,
              "title": "Max Nodes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Graph Graph  Graph Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clusters": {
      "get": {
        "tags": [
          "Graph & Clusters"
        ],
        "summary": "Threat clusters",
        "operationId": "list_clusters_clusters_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Confidence"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "investigation_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Investigation Id"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "updated_at",
                "confidence",
                "created_at"
              ],
              "type": "string",
              "default": "updated_at",
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ClusterListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clusters/{cluster_id}": {
      "get": {
        "tags": [
          "Graph & Clusters"
        ],
        "summary": "Cluster detail",
        "operationId": "get_cluster_clusters__cluster_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Graph & Clusters"
        ],
        "summary": "Rename, tag or re-rate a cluster",
        "operationId": "update_cluster_clusters__cluster_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClusterUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clusters/{cluster_id}/notes": {
      "post": {
        "tags": [
          "Graph & Clusters"
        ],
        "summary": "Add an analyst note to a cluster",
        "operationId": "add_cluster_note_clusters__cluster_id__notes_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClusterNote"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clusters/rebuild": {
      "post": {
        "tags": [
          "Graph & Clusters"
        ],
        "summary": "Re-cluster the entire inventory over distinctive shared fingerprints",
        "operationId": "rebuild_clusters_clusters_rebuild_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Rebuild Clusters Clusters Rebuild Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/settings/scoring": {
      "get": {
        "tags": [
          "Graph & Clusters"
        ],
        "summary": "Correlation scoring weights and thresholds",
        "operationId": "scoring_settings_scoring_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoringConfigOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Graph & Clusters"
        ],
        "summary": "Update scoring weights and thresholds",
        "operationId": "update_scoring_settings_scoring_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoringConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoringConfigOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Graph & Clusters"
        ],
        "summary": "Restore default scoring",
        "operationId": "delete_scoring_settings_scoring_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClusterListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "confidence": {
            "type": "integer",
            "title": "Confidence"
          },
          "confidence_level": {
            "type": "string",
            "title": "Confidence Level"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          },
          "status_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Status Breakdown"
          },
          "brands": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Brands"
          },
          "domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Domains"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "investigation_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Investigation Ids"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "screenshots": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Screenshots"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "confidence",
          "confidence_level",
          "severity",
          "created_at",
          "updated_at"
        ],
        "title": "ClusterListItem"
      },
      "ClusterNote": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ClusterNote"
      },
      "ClusterOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "confidence": {
            "type": "integer",
            "title": "Confidence"
          },
          "confidence_level": {
            "type": "string",
            "title": "Confidence Level"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          },
          "status_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Status Breakdown"
          },
          "brands": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Brands"
          },
          "domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Domains"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "investigation_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Investigation Ids"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "screenshots": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Screenshots"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "asset_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Asset Ids"
          },
          "fingerprint_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fingerprint Ids"
          },
          "ips": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ips"
          },
          "certificates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Certificates"
          },
          "favicons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Favicons"
          },
          "tracking_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tracking Ids"
          },
          "logos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Logos"
          },
          "notes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "confidence",
          "confidence_level",
          "severity",
          "created_at",
          "updated_at"
        ],
        "title": "ClusterOut"
      },
      "ClusterUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(critical|high|medium|low|info)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          }
        },
        "type": "object",
        "title": "ClusterUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Message": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Message"
      },
...
      "Page_ClusterListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ClusterListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[ClusterListItem]"
      },
...
      "ScoringConfig": {
        "properties": {
          "weights": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Weights"
          },
          "thresholds": {
            "$ref": "#/components/schemas/Thresholds"
          },
          "similarity": {
            "$ref": "#/components/schemas/SimilarityThresholds"
          },
          "noisy_fingerprint_limit": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 5.0,
            "title": "Noisy Fingerprint Limit",
            "description": "Fingerprints shared by more assets than this (CDN IPs, default favicons) count at 25% weight",
            "default": 150
          }
        },
        "type": "object",
        "title": "ScoringConfig"
      },
      "ScoringConfigOut": {
        "properties": {
          "weights": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Weights"
          },
          "thresholds": {
            "$ref": "#/components/schemas/Thresholds"
          },
          "similarity": {
            "$ref": "#/components/schemas/SimilarityThresholds"
          },
          "noisy_fingerprint_limit": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 5.0,
            "title": "Noisy Fingerprint Limit",
            "description": "Fingerprints shared by more assets than this (CDN IPs, default favicons) count at 25% weight",
            "default": 150
          },
          "labels": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Labels"
          },
          "defaults": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Defaults"
          }
        },
        "type": "object",
        "title": "ScoringConfigOut"
      },
...
      "Thresholds": {
        "properties": {
          "very_high": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Very High",
            "default": 90
          },
          "high": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "High",
            "default": 70
          },
          "medium": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Medium",
            "default": 50
          },
          "low": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Low",
            "default": 30
          },
          "cluster_min": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Cluster Min",
            "description": "Minimum confidence for cluster membership",
            "default": 50
          }
        },
        "type": "object",
        "title": "Thresholds"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "JWT access token from POST /auth/login",
        "scheme": "bearer"
      }
    }
  },
...
}